Formality result repository.

+++ formality-service/src/main/java/pl/tul/emailsemotions/formalityservice/repositories/FormalityResultRepository.java
@@ -3,6 +3,6 @@ package pl.tul.emailsemotions.formalityservice.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.tul.emailsemotions.formalityservice.model.FormalityResult;
 
-public interface FormalityResultRepository extends JpaRepository<FormalityResult, Integer> {
+public interface FormalityResultRepository extends JpaRepository<FormalityResult, Long> {
 }
